Fixed ft_strncmp.c

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: crenfrow <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/22 19:55:21 by crenfrow          #+#    #+#             */
-/*   Updated: 2016/10/05 13:51:59 by crenfrow         ###   ########.fr       */
+/*   Updated: 2016/10/20 12:38:24 by crenfrow         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,12 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	unsigned char	*dst_tmp;
 	unsigned char	*src_tmp;
-	int				i;
+	unsigned int	i;
 
 	dst_tmp = (unsigned char *)dst;
 	src_tmp = (unsigned char *)src;
 	i = 0;
-	while (i < (int)n)
+	while (i < n)
 	{
 		dst_tmp[i] = src_tmp[i];
 		i++;
